+++ .gitignore
@@ -1,0 +1,2 @@
+Assignment.V3.pdf
+.vs/slnx.sqlite
