+++ function.h
@@ -1,0 +1,36 @@
+/**********|**********|**********|
+Program: YOUR_FILENAME.cpp / YOUR_FILENAME.h
+Course: Data Structures and Algorithms
+Trimester: 2410
+Name: Frank Carrano
+ID: 1071001234
+Lecture Section: TC101
+Tutorial Section: TT1L
+Email: abc123@yourmail.com
+Phone: 018-1234567
+**********|**********|**********/
+
+class Robot
+{
+private:
+public:
+    string file_name = "Robot_config.cfg";
+    string config_info = "";
+};
+
+class movingRobot : public Robot
+{
+};
+
+class shootingRobot : public Robot
+{
+};
+
+class SeeingRobot : public Robot
+{
+};
+class ThinkingRobot : public Robot
+{
+};
+#include <iostream>
+using namespace std;
