+++ main.cpp
@@ -1,31 +1,20 @@
-
+/**********|**********|**********|
+Program: YOUR_FILENAME.cpp / YOUR_FILENAME.h
+Course: Data Structures and Algorithms
+Trimester: 2410
+Name: Frank Carrano
+ID: 1071001234
+Lecture Section: TC101
+Tutorial Section: TT1L
+Email: abc123@yourmail.com
+Phone: 018-1234567
+**********|**********|**********/
 #include <iostream>
 #include <fstream>
 #include <sstream>
 using namespace std;
+#include "function.h"
 
-class Robot
-{
-private:
-public:
-    string file_name = "Robot_config.cfg";
-    string config_info = "";
-};
-
-class movingRobot : public Robot
-{
-};
-
-class shootingRobot : public Robot
-{
-};
-
-class SeeingRobot : public Robot
-{
-};
-class ThinkingRobot : public Robot
-{
-};
 int main()
 {
     system("cd");
@@ -39,6 +28,8 @@ int main()
     {
         cout << "file opening error please check your file name again" << endl;
     }
+    else
+        cout << "config file opened" << endl;
 
     while (getline(file_in, information))
     {
